Address reproducibility issues Fixes #121
