+++ preprocessing/1_lincs.py
@@ -10,30 +10,27 @@
 # ---
 
 # %%
+import os
+import warnings
+
 import numpy as np
 import pandas as pd
 import scanpy as sc
+from scipy import sparse
 from tqdm.notebook import tqdm
+
+from chemCPA.helper import rank_genes_groups_by_cov
+from chemCPA.paths import DATA_DIR
 
 sc.set_figure_params(dpi=100, frameon=False)
 sc.logging.print_header()
-
-# %%
-import os
-
-os.chdir("./../")
-# %%
-import warnings
-
-from chemCPA.helper import rank_genes_groups_by_cov
-
 warnings.filterwarnings("ignore")
 
 # %%
-full = False
+full = True
 load_adata = True
-adata_in = "datasets/lincs_full.h5ad" if full else "datasets/lincs.h5ad"
-adata = sc.read(adata_in) if load_adata else None
+adata_in = "lincs_full.h5ad" if full else "lincs.h5ad"
+adata = sc.read(DATA_DIR / adata_in) if load_adata else None
 
 adata_out = "".join(adata_in.split(".")[:-1]) + "_pp.h5ad"
 adata_out
@@ -43,7 +40,9 @@ adata.obs["condition"] = adata.obs["pert_iname"]
 adata.obs["condition"] = adata.obs["condition"].str.replace("/", "|")
 
 adata.obs["cell_type"] = adata.obs["cell_id"]
-adata.obs["dose_val"] = adata.obs["pert_dose"]
+adata.obs["dose_val"] = adata.obs["pert_dose"].astype(float) / np.max(
+    adata.obs["pert_dose"].astype(float)
+)
 adata.obs["cov_drug_dose_name"] = (
     adata.obs.cell_type.astype(str)
     + "_"
@@ -54,7 +53,7 @@ adata.obs["cov_drug_dose_name"] = (
 adata.obs["cov_drug_name"] = (
     adata.obs.cell_type.astype(str) + "_" + adata.obs.condition.astype(str)
 )
-adata.obs["eval_category"] = adata.obs["cov_drug_name"]
+adata.obs["eval_category"] = adata.obs["cov_drug_dose_name"]
 adata.obs["control"] = (adata.obs["condition"] == "DMSO").astype(int)
 
 # adata.obs['cov_drug_dose_name'] = adata.obs['cov_drug_dose_name'].str.replace('/','|')
@@ -106,13 +105,12 @@ else:
 # %% [markdown]
 # Mapping from `rank_genes_groups_cov` might cause problems when drug contains '_'
 
-
 # %%
 def extract_drug(cond):
     split = cond.split("_")
     if len(split) == 2:
         return split[-1]
-    return "_".join(split[1:])
+    return "_".join(split[1:-1])
 
 
 adata.obs["cov_drug_dose_name"].apply(lambda s: len(s.split("_"))).value_counts()
@@ -156,18 +154,16 @@ except:
 
 # %%
 # code compatibility
-from scipy import sparse
-
 adata.X = sparse.csr_matrix(adata.X)
 
 # %%
-sc.write(adata_out, adata)
+sc.write(DATA_DIR / adata_out, adata)
 
 # %%
 print("all done.")
 
 # %% [markdown]
-# ### Check that `adata.uns[rank_genes_groups_cov]` has all entries in `adata.obs.cov_drug_dose_name` as keys
+# ### Check that `adata.uns[rank_genes_groups_cov]` has all entries in `adata.obs.cov_drug_name` as keys
 
 # %%
 for i, k in enumerate(adata.obs.eval_category.unique()):
@@ -180,7 +176,7 @@ for i, k in enumerate(adata.obs.eval_category.unique()):
 # ### Checking the same for the stored adata object
 
 # %%
-adata_2 = sc.read(adata_out)
+adata_2 = sc.read(DATA_DIR / adata_out)
 
 # %%
 for i, k in enumerate(adata_2.obs.eval_category.unique()):
@@ -198,5 +194,11 @@ set(list(adata.uns["rank_genes_groups_cov"])) - set(
 set((list(adata_2.uns["rank_genes_groups_cov"]))) - set(
     list(adata.uns["rank_genes_groups_cov"])
 )
+
+# %%
+len(list(adata_2.uns["rank_genes_groups_cov"].keys()))
+
+# %%
+adata.obs["dose_val"].value_counts()
 
 # %%
